+++ src/RouterMatch.php
@@ -24,11 +24,25 @@ class RouterMatch
      */
     public $format;
 
-    public function __construct($controller, $action, $args, $format)
+    /**
+     * User-defined options set on the route.
+     *
+     * In practice this is all the options passed to ``addRroute`` except:
+     *
+     * - ``_format``
+     * - ``as``
+     * - ``except``
+     * - ``only``
+     * - ``to``
+     */
+    public $options;
+
+    public function __construct($controller, $action, $args, $format, array $options)
     {
         $this->controller = $controller;
         $this->action = $action;
         $this->args = $args;
         $this->format = $format;
+        $this->options = $options;
     }
 }